Moved identical logic from desktop and browser into system.service
